vernac/declaremods: make object collection tail-recursive
This patch is trying to upstream a jsCoq patch that was written to
solve Stack Overflow issues:

  https://github.com/jscoq/jscoq/blob/v8.12/etc/patches/cps.patch

It turns List.fold_right in vernac/declaremods into List.fold_left
followed by List.rev (the output of the fold is a list which is now
constructed in reverse order).

My *naive* understanding of the code is that the *order* in which the
collection is performed should *not* be observable, so that the switch
to List.fold_right to List.fold_left should be correct. In particular,
notice that some slight effects (calling Lib.make_oname) were called
from right to left in the collect_* functions, and then from left to
right in the open_* functions.

Of course, reality differs: this modification appears to change
parameter resolution in an obscure use of a `pow_N` function in
Ncring_tac.v. The existing code would break, and I now have to pass
the (previously implicit) parameters explicitly.

I am not sure whether it means that the current change should be
thrown in the garbage bin (it breaks code), or that there is a bug
somewhere else (sounds somewhat likely to me).